Refactor the unit tests
